django/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
boto3==1.16.25
botocore==1.19.63
certifi==2021.5.30
cffi==1.15.1
chardet==3.0.4
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.1.3
Django==3.1
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.1
django-allauth==0.45.0
django-import-export==2.4.0
django-rest-auth==0.9.5
django-ses==1.0.3
django-storages==1.10.1
django-summernote==0.8.11.6
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
drfpasswordless==1.5.7
et-xmlfile==1.1.0
fabric==2.6.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
invoke==1.6.0
jmespath==0.10.0
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
paramiko==2.7.2
pathlib2==2.3.6
Pillow==9.2.0
plotly==4.14.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
twilio==6.61.0
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
django-filter==21.1
